# feat(performance-monitoring): implement real-time performance monitoring dashboard