… PERFORMANCE_MONITORING_IMPLEMENTATION.md
… src/js/components/performance-monitoring-dashboard.js
… src/main.js
… tests/unit/performance-monitoring-validation.test.js
… tests/unit/performance-monitoring.test.js